Added lightning effect added headline1.png fixed some bugs

+++ app/src/main/java/de/traendy/spaceshooter/obstacle/MeteorEntityHolder.kt
@@ -1,6 +1,7 @@
 package de.traendy.spaceshooter.obstacle
 
 import android.graphics.Canvas
+import de.traendy.spaceshooter.effects.Lightning
 import de.traendy.spaceshooter.engine.CollisionDetector
 import de.traendy.spaceshooter.engine.Entity
 import de.traendy.spaceshooter.engine.PrimitiveEntityHolder
@@ -28,7 +29,7 @@ class MeteorEntityHolder(
         }
     }
 
-    fun updateMeteors(dangerousEntities: List<Entity>, player: Player?, canvas: Canvas) {
+    fun updateMeteors(dangerousEntities: List<Entity>, player: Player?, canvas: Canvas, damageEffect: Lightning) {
         getAllEntities().forEach { meteor ->
             detectDestruction(dangerousEntities, meteor)
             if (!meteor.isAlive()) {
@@ -36,19 +37,21 @@ class MeteorEntityHolder(
             } else {
                 meteor.updatePosition(0f, 0f)
                 meteor.draw(canvas)
-                detectPlayerCollision(player, meteor)
+                detectPlayerCollision(player, meteor, damageEffect)
             }
         }
     }
 
     private fun detectPlayerCollision(
         player: Player?,
-        meteor: Meteor
+        meteor: Meteor,
+        damageEffect: Lightning
     ) {
         player?.let {
-            if (collisionDetector.collided(meteor, player) && meteor.isAlive()) {
+            if (collisionDetector.collided(meteor, player) && meteor.isAlive() && player.isAlive()) {
                 meteor.kill()
                 player.kill()
+                damageEffect.show()
             }
         }
     }
